fix(31-09): width-reconcile genuine class-(B) members unmasked by 31-08 (BITS-02)
- CuiCombatManager::fillSpamOrder: pos unsigned int -> size_t to bind getFirstToken's size_t& endpos on x64 (parse cursor, not serialized)
- MeshConstructionHelper: 5x VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE 0U -> static_cast<size_t>(0) so the template deduces one T against size_t index/.size() args (C2672)
- TcpClient/TcpServer: IOCP completionKey unsigned long int -> ULONG_PTR for GetQueuedCompletionStatus's PULONG_PTR arg (C2664); local IOCP key only, not on the wire
- no on-the-wire/serialized field width changed (only local length/count/cursor args)

+++ src/engine/client/library/clientUserInterface/src/shared/core/CuiCombatManager.cpp
@@ -1888,7 +1888,7 @@ void CuiCombatManager::processCombatSpam (const MessageQueueCombatSpam & spamMsg
 
 void CuiCombatManager::fillSpamOrder(short spamOrder[], Unicode::String spamTemplate)
 {
-	unsigned int pos = 0;
+	size_t pos = 0; // parse cursor (not serialized); size_t to bind getFirstToken's size_t& endpos on x64 (BITS-02 width)
 	Unicode::NarrowString token;
 	Unicode::NarrowString narrowTemplate = Unicode::wideToNarrow(spamTemplate);
 	for (int i = 0; pos != narrowTemplate.npos; i++)
